Approved-by: Simon Kuenzer <simon@unikraft.io>

+++ internal/cli/kraft/cloud/volume/import/import.go
@@ -35,8 +35,11 @@ type ImportOptions struct {
 	Token string             `noattribute:"true"`
 	Metro string             `noattribute:"true"`
 
-	Source string `local:"true" long:"source" short:"s" usage:"Path to the data source (directory, Dockerfile)" default:"."`
-	VolID  string `local:"true" long:"volume" short:"v" usage:"Identifier of an existing volume (name or UUID)"`
+	VolimportImage string `local:"true" long:"image" usage:"Volume import image to use" default:"utils/volimport:latest"`
+	Force          bool   `local:"true" long:"force" short:"f" usage:"Force import, even if it might fail"`
+	Source         string `local:"true" long:"source" short:"s" usage:"Path to the data source (directory, Dockerfile, Docker link, cpio file)" default:"."`
+	Timeout        uint64 `local:"true" long:"timeout" short:"t" usage:"Timeout for the import process in seconds"`
+	VolID          string `local:"true" long:"volume" short:"v" usage:"Identifier of an existing volume (name or UUID)"`
 }
 
 const volimportPort uint16 = 42069
@@ -49,6 +52,15 @@ func NewCmd() *cobra.Command {
 		Example: heredoc.Doc(`
 			# Import data from a local directory "path/to/data" to a volume named "my-volume"
 			$ kraft cloud volume import --source path/to/data --volume my-volume
+
+			# Import data from a local Dockerfile "path/to/Dockerfile" to a volume named "my-volume"
+			$ kraft cloud volume import --source path/to/Dockerfile --volume my-volume
+
+			# Import data from a docker registry "docker.io/nginx:latest" to a volume named "my-volume"
+			$ kraft cloud volume import --source docker.io/nginx:latest --volume my-volume
+
+			# Import data from a local cpio file "path/to/file" to a volume named "my-volume"
+			$ kraft cloud volume import --source path/to/file --volume my-volume
 		`),
 		Annotations: map[string]string{
 			cmdfactory.AnnotationHelpGroup: "kraftcloud-vol",
@@ -64,6 +76,10 @@ func NewCmd() *cobra.Command {
 func (opts *ImportOptions) Pre(cmd *cobra.Command, _ []string) error {
 	if opts.VolID == "" {
 		return fmt.Errorf("must specify a value for the --volume flag")
+	}
+
+	if finfo, err := os.Stat(opts.Source); err == nil && !finfo.IsDir() {
+		return fmt.Errorf("local source path must be a directory")
 	}
 
 	err := utils.PopulateMetroToken(cmd, &opts.Metro, &opts.Token)
@@ -119,16 +135,17 @@ func importVolumeData(ctx context.Context, opts *ImportOptions) (retErr error) {
 	}
 
 	defer func() {
-		err := os.Remove(cpioPath)
-		if err != nil {
-			err = fmt.Errorf("removing temp CPIO archive: %w", err)
+		if cpioPath != opts.Source {
+			err := os.Remove(cpioPath)
+			if err != nil {
+				err = fmt.Errorf("removing temp CPIO archive: %w", err)
+			}
+			retErr = errors.Join(retErr, err)
 		}
-		retErr = errors.Join(retErr, err)
 	}()
 
 	var volUUID string
-	var volSize int64
-	if volUUID, volSize, err = volumeSanityCheck(ctx, vcli, opts.VolID, cpioSize); err != nil {
+	if volUUID, _, err = volumeSanityCheck(ctx, vcli, opts.VolID, cpioSize); err != nil {
 		return err
 	}
 
@@ -141,7 +158,7 @@ func importVolumeData(ctx context.Context, opts *ImportOptions) (retErr error) {
 			if authStr, err = genRandAuth(); err != nil {
 				return fmt.Errorf("generating random authentication string: %w", err)
 			}
-			instID, instFQDN, err = runVolimport(ctx, icli, volUUID, authStr)
+			instID, instFQDN, err = runVolimport(ctx, icli, opts.VolimportImage, volUUID, authStr, opts.Timeout)
 			return err
 		},
 	)
@@ -163,21 +180,21 @@ func importVolumeData(ctx context.Context, opts *ImportOptions) (retErr error) {
 	// nil error upon context cancellation. We temporarily handle potential copy
 	// errors ourselves here.
 	var copyCPIOErr error
+	var freeSpace uint64
+	var totalSpace uint64
 
 	paraprogress, err := paraProgress(ctx, fmt.Sprintf("Importing data (%s)", humanize.IBytes(uint64(cpioSize))),
 		func(ctx context.Context, callback func(float64)) (retErr error) {
 			instAddr := instFQDN + ":" + strconv.FormatUint(uint64(volimportPort), 10)
 			conn, err := tls.Dial("tcp4", instAddr, nil)
 			if err != nil {
-				return fmt.Errorf("connecting to volume data import instance: %w", err)
+				return fmt.Errorf("connecting to volume data import instance send port: %w", err)
 			}
-			defer func() {
-				retErr = errors.Join(retErr, conn.Close())
-			}()
+			defer conn.Close()
 
 			ctx, cancel := context.WithCancel(ctx)
 			defer cancel()
-			err = copyCPIO(ctx, conn, authStr, cpioPath, cpioSize, callback)
+			freeSpace, totalSpace, err = copyCPIO(ctx, conn, authStr, cpioPath, opts.Force, opts.Timeout, uint64(cpioSize), callback)
 			copyCPIOErr = err
 			return err
 		},
@@ -198,12 +215,12 @@ func importVolumeData(ctx context.Context, opts *ImportOptions) (retErr error) {
 			Value: opts.VolID,
 		},
 		fancymap.FancyMapEntry{
-			Key:   "imported",
-			Value: humanize.IBytes(uint64(cpioSize)),
+			Key:   "free",
+			Value: humanize.IBytes(freeSpace),
 		},
 		fancymap.FancyMapEntry{
-			Key:   "capacity",
-			Value: humanize.IBytes(uint64(volSize)),
+			Key:   "total",
+			Value: humanize.IBytes(totalSpace),
 		},
 	)
